# blocks: split macro into multiple functions, fast!