--- a/appveyor.yml
+++ b/appveyor.yml
@@ -21,7 +21,7 @@ build_script:
 
     appveyor AddMessage "Platform rust: %RUST_INSTALL%"
 
-    appveyor DownloadFile "https://static.rust-lang.org/dist/rust-nightly-%RUST_INSTALL%.exe" -FileName rust-install.exe
+    appveyor DownloadFile "https://static.rust-lang.org/dist/rust-1.40.0-%RUST_INSTALL%.exe" -FileName rust-install.exe
 
     "./rust-install.exe" /VERYSILENT /NORESTART /DIR="C:\Rust\"