@iceiix 2020-01-10
https://users.rust-lang.org/t/5-hours-to-compile-macro-what-can-i-do/36508/3