@iceiix 2020-01-10
@woodgear If you are still interested in building this project, I believe this pull request should address the out of memory errors you were seeing compiling steven_blocks on Windows (#184), finally