@iceiix 2020-01-23
https://old.reddit.com/r/rust/comments/er1xkh/how_refactoring_a_macro_reduced_my_rust_project/
https://news.ycombinator.com/item?id=22093930